refactor: restructure src/ into functional subdirectories
