+++ src/endpoint/srv6_endpoint.h
@@ -1,0 +1,10 @@
+#ifndef SRV6_ENDPOINT_H
+#define SRV6_ENDPOINT_H
+
+// Facade header: includes all endpoint sub-modules
+#include "endpoint/srv6_endpoint_core.h"
+#include "endpoint/srv6_endpoint_basic.h"
+#include "endpoint/srv6_endpoint_decap.h"
+#include "endpoint/srv6_endpoint_l2.h"
+
+#endif // SRV6_ENDPOINT_H
