+++ src/headend/srv6_encaps_red.h
@@ -8,10 +8,10 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_endian.h>
 
-#include "xdp_prog.h"
-#include "srv6.h"
-#include "srv6_headend_utils.h"
-#include "srv6_fib.h"
+#include "core/xdp_prog.h"
+#include "core/srv6.h"
+#include "headend/srv6_headend_utils.h"
+#include "core/srv6_fib.h"
 
 // H.Encaps.Red single-segment: no SRH, just outer IPv6 with nexthdr = inner_proto
 static __always_inline int do_h_encaps_red_1seg(
