+++ src/headend/srv6_insert.h
@@ -8,10 +8,10 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_endian.h>
 
-#include "xdp_prog.h"
-#include "srv6.h"
-#include "srv6_headend_utils.h"
-#include "srv6_fib.h"
+#include "core/xdp_prog.h"
+#include "core/srv6.h"
+#include "headend/srv6_headend_utils.h"
+#include "core/srv6_fib.h"
 
 // H.Insert core implementation (RFC 8986 Section 4.1)
 //
