--- src/srv6_endpoint.h
@@ -1,784 +1,0 @@
-#ifndef SRV6_ENDPOINT_H
-#define SRV6_ENDPOINT_H
-
-#include <linux/types.h>
-#include <linux/if_ether.h>
-#include <linux/ipv6.h>
-#include <linux/in.h>
-#include <bpf/bpf_helpers.h>
-
-#include "xdp_prog.h"
-#include "srv6.h"
-#include "srv6_headend_utils.h"
-#include "srv6_fib.h"
-#include "srv6_decaps.h"
-#include "xdp_stats.h"
-#include "xdp_map.h"
-
-// Endpoint processing context - shared by all endpoint functions
-struct endpoint_ctx {
-    struct xdp_md *ctx;
-    struct ipv6hdr *ip6h;
-    struct ipv6_sr_hdr *srh;
-    struct sid_function_entry *entry;
-    void *data_end;
-    __u8 segments_left;
-    __u8 new_sl;
-};
-
-// Initialize endpoint context and perform common SL checks
-// Returns:
-//   0: Success, context initialized
-//  -1: SL=0, pass to upper layer (not an error)
-//  -2: Invalid SL, should drop
-static __always_inline int endpoint_init(
-    struct endpoint_ctx *ectx,
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    ectx->ctx = ctx;
-    ectx->ip6h = ip6h;
-    ectx->srh = srh;
-    ectx->entry = entry;
-    ectx->data_end = (void *)(long)ctx->data_end;
-    ectx->segments_left = srh->segments_left;
-
-    // RFC 8986: If SL=0, pass to upper layer
-    if (ectx->segments_left == 0) {
-        return -1;
-    }
-
-    // Decrement SL
-    ectx->new_sl = ectx->segments_left - 1;
-
-    // Verify bounds
-    if (ectx->new_sl > srh->first_segment || ectx->new_sl > 9) {
-        return -2;
-    }
-
-    return 0;
-}
-
-// Update Destination Address with Segment List[new_sl]
-// Common operation for End, End.X, End.T, etc.
-static __always_inline int endpoint_update_da(struct endpoint_ctx *ectx)
-{
-    void *seg_base = (void *)ectx->srh + 8;
-    if (copy_segment_by_index(&ectx->ip6h->daddr, seg_base, ectx->data_end, ectx->new_sl) != 0) {
-        return -1;
-    }
-    ectx->srh->segments_left = ectx->new_sl;
-    return 0;
-}
-
-// Core FIB lookup + redirect with explicit dst address
-// Re-derives all packet pointers from ctx->data (safe after bpf_xdp_adjust_head)
-static __always_inline int endpoint_fib_redirect_core(struct endpoint_ctx *ectx, void *dst, __u32 fib_ifindex)
-{
-    void *data = (void *)(long)ectx->ctx->data;
-    void *data_end = (void *)(long)ectx->ctx->data_end;
-    struct ethhdr *eth = data;
-    if ((void *)eth + sizeof(*eth) > data_end) {
-        return XDP_DROP;
-    }
-
-    struct ipv6hdr *ip6h = (struct ipv6hdr *)(eth + 1);
-    if ((void *)(ip6h + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    __u32 ifindex;
-    int fib_result = srv6_fib_lookup_v6_core(ectx->ctx, ip6h, eth, &ifindex, dst, fib_ifindex);
-
-    switch (fib_result) {
-    case FIB_RESULT_REDIRECT:
-        STATS_INC(STATS_SRV6_END, 0);
-        return bpf_redirect(ifindex, 0);
-    case FIB_RESULT_DROP:
-        return XDP_DROP;
-    default:
-        return XDP_PASS;
-    }
-}
-
-// FIB lookup on ip6h->daddr (for End, End.T)
-static __always_inline int endpoint_fib_redirect(struct endpoint_ctx *ectx, __u32 fib_ifindex)
-{
-    void *data = (void *)(long)ectx->ctx->data;
-    void *data_end = (void *)(long)ectx->ctx->data_end;
-    struct ipv6hdr *ip6h = (void *)data + ETH_HLEN;
-    if ((void *)(ip6h + 1) > data_end) {
-        return XDP_DROP;
-    }
-    return endpoint_fib_redirect_core(ectx, &ip6h->daddr, fib_ifindex);
-}
-
-// FIB lookup on entry->nexthop (for End.X)
-static __always_inline int endpoint_fib_redirect_nexthop(struct endpoint_ctx *ectx)
-{
-    return endpoint_fib_redirect_core(ectx, ectx->entry->nexthop, ectx->ctx->ingress_ifindex);
-}
-
-// ========================================================================
-// SRv6 Flavor Helpers (PSP, USP, USD)
-// ========================================================================
-
-// Strip SRH from packet, preserving Ethernet + IPv6 headers
-// Uses the same adjust_head pattern as srv6_decap():
-//   1. Save Eth + IPv6 header to stack
-//   2. Strip all (Eth + IPv6 + SRH) via adjust_head(+)
-//   3. Re-expand Eth + IPv6 via adjust_head(-)
-//   4. Restore saved headers with updated nexthdr/payload_len
-//
-// Before: [Eth(14)][IPv6(40)][SRH][Upper Layer]
-// After:  [Eth(14)][IPv6(40)][Upper Layer]
-//
-// Returns: 0 on success, -1 on failure
-// Note: After success, caller must re-fetch all pointers from ctx->data
-static __always_inline int endpoint_strip_srh(struct endpoint_ctx *ectx)
-{
-    void *data = (void *)(long)ectx->ctx->data;
-    void *data_end = (void *)(long)ectx->ctx->data_end;
-
-    struct ethhdr *eth = data;
-    if ((void *)(eth + 1) > data_end)
-        return -1;
-
-    struct ipv6hdr *ip6h = (struct ipv6hdr *)(eth + 1);
-    if ((void *)(ip6h + 1) > data_end)
-        return -1;
-
-    // Save headers to stack (54 bytes total, well within 512B BPF stack limit)
-    struct ethhdr saved_eth;
-    struct ipv6hdr saved_ip6h;
-    __builtin_memcpy(&saved_eth, eth, sizeof(saved_eth));
-    __builtin_memcpy(&saved_ip6h, ip6h, sizeof(saved_ip6h));
-
-    // Save SRH metadata before stripping
-    __u8 inner_nexthdr = ectx->srh->nexthdr;
-    int srh_len = 8 + (ectx->srh->hdrlen * 8);
-
-    // Strip Eth + IPv6 + SRH
-    int total_strip = ETH_HLEN + sizeof(struct ipv6hdr) + srh_len;
-    if (bpf_xdp_adjust_head(ectx->ctx, total_strip))
-        return -1;
-
-    // Re-expand for Eth + IPv6 header
-    if (bpf_xdp_adjust_head(ectx->ctx, -(int)(ETH_HLEN + sizeof(struct ipv6hdr))))
-        return -1;
-
-    // Re-fetch pointers
-    data = (void *)(long)ectx->ctx->data;
-    data_end = (void *)(long)ectx->ctx->data_end;
-
-    eth = data;
-    if ((void *)(eth + 1) > data_end)
-        return -1;
-
-    ip6h = (struct ipv6hdr *)(eth + 1);
-    if ((void *)(ip6h + 1) > data_end)
-        return -1;
-
-    // Restore saved headers
-    __builtin_memcpy(eth, &saved_eth, sizeof(saved_eth));
-    __builtin_memcpy(ip6h, &saved_ip6h, sizeof(saved_ip6h));
-
-    // Update IPv6 header: point to upper layer, adjust length
-    ip6h->nexthdr = inner_nexthdr;
-    ip6h->payload_len = bpf_htons(bpf_ntohs(ip6h->payload_len) - srh_len);
-
-    return 0;
-}
-
-// Handle USD flavor: decapsulate inner packet at SL=0
-// Performs full decap (strip Eth+IPv6+SRH) and FIB redirect on inner packet
-static __always_inline int endpoint_handle_usd(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    __u8 inner_proto = srh->nexthdr;
-
-    if (inner_proto == IPPROTO_IPIP) {
-        DEBUG_PRINT("USD: Decapsulating inner IPv4\n");
-        if (srv6_decap(ctx, srh, IPPROTO_IPIP) != 0)
-            return XDP_DROP;
-
-        void *data = (void *)(long)ctx->data;
-        void *data_end = (void *)(long)ctx->data_end;
-        struct ethhdr *eth = data;
-        if ((void *)(eth + 1) > data_end) return XDP_DROP;
-        struct iphdr *iph = (void *)(eth + 1);
-        if ((void *)(iph + 1) > data_end) return XDP_DROP;
-        eth->h_proto = bpf_htons(ETH_P_IP);
-        STATS_INC(STATS_SRV6_END, 0);
-        int action = srv6_fib_redirect_v4(ctx, iph, eth, ctx->ingress_ifindex);
-        return (action == XDP_PASS) ? XDP_DROP : action;
-    }
-
-    if (inner_proto == IPPROTO_IPV6) {
-        DEBUG_PRINT("USD: Decapsulating inner IPv6\n");
-        if (srv6_decap(ctx, srh, IPPROTO_IPV6) != 0)
-            return XDP_DROP;
-
-        void *data = (void *)(long)ctx->data;
-        void *data_end = (void *)(long)ctx->data_end;
-        struct ethhdr *eth = data;
-        if ((void *)(eth + 1) > data_end) return XDP_DROP;
-        struct ipv6hdr *inner_ip6h = (void *)(eth + 1);
-        if ((void *)(inner_ip6h + 1) > data_end) return XDP_DROP;
-        STATS_INC(STATS_SRV6_END, 0);
-        int action = srv6_fib_redirect(ctx, inner_ip6h, eth, ctx->ingress_ifindex);
-        return (action == XDP_PASS) ? XDP_DROP : action;
-    }
-
-    DEBUG_PRINT("USD: Unsupported inner protocol %d\n", inner_proto);
-    return XDP_DROP;
-}
-
-// Handle USP flavor: strip SRH at SL=0 and perform FIB lookup
-// Called when SL=0 and USP flavor is set
-static __always_inline int endpoint_handle_usp(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry,
-    __u32 fib_ifindex)
-{
-    struct endpoint_ctx ectx;
-    ectx.ctx = ctx;
-    ectx.ip6h = ip6h;
-    ectx.srh = srh;
-    ectx.entry = entry;
-    ectx.data_end = (void *)(long)ctx->data_end;
-
-    if (endpoint_strip_srh(&ectx) != 0) {
-        DEBUG_PRINT("USP: Failed to strip SRH\n");
-        return XDP_DROP;
-    }
-
-    DEBUG_PRINT("USP: Stripped SRH, FIB lookup on DA\n");
-    return endpoint_fib_redirect(&ectx, fib_ifindex);
-}
-
-// ========================================================================
-// Endpoint Function Implementations
-// ========================================================================
-
-// End: Basic SRv6 endpoint operation
-// RFC 8986 Section 4.1
-// Supports flavors: PSP (strip SRH at penultimate), USP (strip SRH at ultimate),
-//                   USD (decapsulate at ultimate)
-static __always_inline int process_end(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    struct endpoint_ctx ectx;
-    int ret = endpoint_init(&ectx, ctx, ip6h, srh, entry);
-
-    if (ret == -1) {
-        // SL=0: check USD and USP flavors
-        if (entry->flavor == SRV6_LOCAL_FLAVOR_USD) {
-            return endpoint_handle_usd(ctx, ip6h, srh, entry);
-        } else if (entry->flavor == SRV6_LOCAL_FLAVOR_USP) {
-            return endpoint_handle_usp(ctx, ip6h, srh, entry, ctx->ingress_ifindex);
-        }
-        DEBUG_PRINT("End: SL is 0, passing to upper layer\n");
-        return XDP_PASS;
-    }
-    if (ret == -2) {
-        DEBUG_PRINT("End: Invalid SL\n");
-        return XDP_DROP;
-    }
-
-    if (endpoint_update_da(&ectx) != 0) {
-        DEBUG_PRINT("End: Failed to update DA\n");
-        return XDP_DROP;
-    }
-
-    // PSP: strip SRH when new_sl becomes 0 (penultimate segment)
-    if (ectx.new_sl == 0 && (entry->flavor == SRV6_LOCAL_FLAVOR_PSP)) {
-        DEBUG_PRINT("End+PSP: Stripping SRH at penultimate\n");
-        if (endpoint_strip_srh(&ectx) != 0) {
-            return XDP_DROP;
-        }
-    }
-
-    DEBUG_PRINT("End: Updated DA, new SL=%d\n", ectx.new_sl);
-    return endpoint_fib_redirect(&ectx, ectx.ctx->ingress_ifindex);
-}
-
-// End.X: Layer-3 cross-connect to specified nexthop
-// RFC 8986 Section 4.2
-// Like End but uses entry->nexthop for FIB lookup instead of the updated DA
-// Supports flavors: PSP, USP, USD
-static __always_inline int process_end_x(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    struct endpoint_ctx ectx;
-    int ret = endpoint_init(&ectx, ctx, ip6h, srh, entry);
-    if (ret == -1) {
-        if (entry->flavor == SRV6_LOCAL_FLAVOR_USD) {
-            return endpoint_handle_usd(ctx, ip6h, srh, entry);
-        } else if (entry->flavor == SRV6_LOCAL_FLAVOR_USP) {
-            return endpoint_handle_usp(ctx, ip6h, srh, entry, ctx->ingress_ifindex);
-        }
-        return XDP_PASS;
-    }
-    if (ret == -2) return XDP_DROP;
-
-    if (endpoint_update_da(&ectx) != 0) {
-        return XDP_DROP;
-    }
-
-    if (ectx.new_sl == 0 && (entry->flavor == SRV6_LOCAL_FLAVOR_PSP)) {
-        DEBUG_PRINT("End.X+PSP: Stripping SRH at penultimate\n");
-        if (endpoint_strip_srh(&ectx) != 0) {
-            return XDP_DROP;
-        }
-    }
-
-    DEBUG_PRINT("End.X: Updated DA, new SL=%d, forwarding via nexthop\n", ectx.new_sl);
-    return endpoint_fib_redirect_nexthop(&ectx);
-}
-
-// End.T: Lookup in specific routing table
-// RFC 8986 Section 4.3
-// Like End but uses entry->vrf_ifindex for VRF-aware FIB lookup
-// Supports flavors: PSP, USP, USD
-static __always_inline int process_end_t(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    __u32 fib_ifindex = entry->vrf_ifindex ? entry->vrf_ifindex : ctx->ingress_ifindex;
-
-    struct endpoint_ctx ectx;
-    int ret = endpoint_init(&ectx, ctx, ip6h, srh, entry);
-    if (ret == -1) {
-        if (entry->flavor == SRV6_LOCAL_FLAVOR_USD) {
-            return endpoint_handle_usd(ctx, ip6h, srh, entry);
-        } else if (entry->flavor == SRV6_LOCAL_FLAVOR_USP) {
-            return endpoint_handle_usp(ctx, ip6h, srh, entry, fib_ifindex);
-        }
-        return XDP_PASS;
-    }
-    if (ret == -2) return XDP_DROP;
-
-    if (endpoint_update_da(&ectx) != 0) {
-        return XDP_DROP;
-    }
-
-    if (ectx.new_sl == 0 && (entry->flavor == SRV6_LOCAL_FLAVOR_PSP)) {
-        DEBUG_PRINT("End.T+PSP: Stripping SRH at penultimate\n");
-        if (endpoint_strip_srh(&ectx) != 0) {
-            return XDP_DROP;
-        }
-    }
-
-    DEBUG_PRINT("End.T: Updated DA, new SL=%d, FIB ifindex=%d\n", ectx.new_sl, fib_ifindex);
-    return endpoint_fib_redirect(&ectx, fib_ifindex);
-}
-
-// End.DX4: Decapsulation with IPv4 cross-connect
-// RFC 8986 Section 4.6
-static __always_inline int process_end_dx4(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    // 1. RFC 8986: SL must be 0 for DX4
-    if (srh->segments_left != 0) {
-        DEBUG_PRINT("End.DX4: SL != 0, passing\n");
-        return XDP_PASS;
-    }
-
-    // 2. Strip outer IPv6+SRH, expose inner IPv4
-    if (srv6_decap(ctx, srh, IPPROTO_IPIP) != 0) {
-        DEBUG_PRINT("End.DX4: Decapsulation failed\n");
-        return XDP_DROP;
-    }
-
-    // 3. Re-fetch pointers after adjust_head
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-
-    // 4. Validate Ethernet + IPv4 headers
-    struct ethhdr *eth = data;
-    if ((void *)(eth + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    struct iphdr *iph = (void *)(eth + 1);
-    if ((void *)(iph + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    // 5. Set EtherType to IPv4
-    eth->h_proto = bpf_htons(ETH_P_IP);
-
-    // 6. FIB lookup on inner IPv4 and redirect
-    DEBUG_PRINT("End.DX4: Decapsulated, forwarding inner IPv4\n");
-    STATS_INC(STATS_SRV6_END, 0);
-
-    // After decapsulation, we must not return XDP_PASS because:
-    // 1. Packet structure has changed (IPv4 instead of IPv6+SRH)
-    // 2. Old pointers in caller are invalidated
-    // So if FIB lookup fails (XDP_PASS), convert to XDP_DROP
-    int action = srv6_fib_redirect_v4(ctx, iph, eth, ctx->ingress_ifindex);
-    if (action == XDP_PASS) {
-        DEBUG_PRINT("End.DX4: FIB lookup failed, dropping\n");
-        return XDP_DROP;
-    }
-    return action;
-}
-
-// End.DX6: Decapsulation with IPv6 cross-connect
-// RFC 8986 Section 4.5
-static __always_inline int process_end_dx6(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    // 1. RFC 8986: SL must be 0 for DX6
-    if (srh->segments_left != 0) {
-        DEBUG_PRINT("End.DX6: SL != 0, passing\n");
-        return XDP_PASS;
-    }
-
-    // 2. Strip outer IPv6+SRH, expose inner IPv6
-    if (srv6_decap(ctx, srh, IPPROTO_IPV6) != 0) {
-        DEBUG_PRINT("End.DX6: Decapsulation failed\n");
-        return XDP_DROP;
-    }
-
-    // 3. Re-fetch pointers after adjust_head
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-
-    // 4. Validate Ethernet + IPv6 headers
-    struct ethhdr *eth = data;
-    if ((void *)(eth + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    struct ipv6hdr *inner_ip6h = (void *)(eth + 1);
-    if ((void *)(inner_ip6h + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    // 5. EtherType is already IPv6 (unchanged)
-
-    // 6. FIB lookup on inner IPv6 and redirect
-    DEBUG_PRINT("End.DX6: Decapsulated, forwarding inner IPv6\n");
-    STATS_INC(STATS_SRV6_END, 0);
-
-    // After decapsulation, we must not return XDP_PASS because:
-    // 1. Packet structure has changed (inner IPv6 only, outer IPv6+SRH stripped)
-    // 2. Old pointers in caller are invalidated
-    // So if FIB lookup fails (XDP_PASS), convert to XDP_DROP
-    int action = srv6_fib_redirect(ctx, inner_ip6h, eth, ctx->ingress_ifindex);
-    if (action == XDP_PASS) {
-        DEBUG_PRINT("End.DX6: FIB lookup failed, dropping\n");
-        return XDP_DROP;
-    }
-    return action;
-}
-
-// End.DX2: Decapsulation with L2 cross-connect
-// RFC 8986 Section 4.10
-// Strips outer Ethernet + IPv6 + SRH, exposes inner L2 frame,
-// and redirects to a specified output interface (OIF).
-// OIF is stored as __u32 in the first 4 bytes of entry->nexthop.
-static __always_inline int process_end_dx2(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    if (srh->segments_left != 0) {
-        DEBUG_PRINT("End.DX2: SL != 0, passing\n");
-        return XDP_PASS;
-    }
-
-    if (srh->nexthdr != IPPROTO_ETHERNET) {
-        DEBUG_PRINT("End.DX2: nexthdr is not Ethernet (%d)\n", srh->nexthdr);
-        return XDP_DROP;
-    }
-
-    __u32 oif;
-    __builtin_memcpy(&oif, entry->nexthop, sizeof(__u32));
-    if (oif == 0) {
-        DEBUG_PRINT("End.DX2: OIF not configured\n");
-        return XDP_DROP;
-    }
-
-    int strip_len = calc_decap_strip_len(srh);
-    if (bpf_xdp_adjust_head(ctx, strip_len)) {
-        DEBUG_PRINT("End.DX2: bpf_xdp_adjust_head failed\n");
-        return XDP_DROP;
-    }
-
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-    struct ethhdr *inner_eth = data;
-    if ((void *)(inner_eth + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    DEBUG_PRINT("End.DX2: Decapsulated, redirect to ifindex %d\n", oif);
-    STATS_INC(STATS_SRV6_END, 0);
-    return bpf_redirect(oif, 0);
-}
-
-// End.DT4: Decapsulation with IPv4 table lookup
-// RFC 8986 Section 4.8
-// Like End.DX4 but uses VRF-aware FIB lookup via entry->vrf_ifindex
-static __always_inline int process_end_dt4(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    // 1. RFC 8986: SL must be 0 for DT4
-    if (srh->segments_left != 0) {
-        DEBUG_PRINT("End.DT4: SL != 0, passing\n");
-        return XDP_PASS;
-    }
-
-    // 2. Strip outer IPv6+SRH, expose inner IPv4
-    if (srv6_decap(ctx, srh, IPPROTO_IPIP) != 0) {
-        DEBUG_PRINT("End.DT4: Decapsulation failed\n");
-        return XDP_DROP;
-    }
-
-    // 3. Re-fetch pointers after adjust_head
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-
-    // 4. Validate Ethernet + IPv4 headers
-    struct ethhdr *eth = data;
-    if ((void *)(eth + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    struct iphdr *iph = (void *)(eth + 1);
-    if ((void *)(iph + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    // 5. Set EtherType to IPv4
-    eth->h_proto = bpf_htons(ETH_P_IP);
-
-    // 6. FIB lookup with VRF ifindex (0 falls back to ingress ifindex)
-    __u32 fib_ifindex = entry->vrf_ifindex ? entry->vrf_ifindex : ctx->ingress_ifindex;
-    DEBUG_PRINT("End.DT4: Decapsulated, FIB ifindex=%d\n", fib_ifindex);
-    STATS_INC(STATS_SRV6_END, 0);
-
-    int action = srv6_fib_redirect_v4(ctx, iph, eth, fib_ifindex);
-    // After decapsulation, must not return XDP_PASS
-    if (action == XDP_PASS) {
-        DEBUG_PRINT("End.DT4: FIB lookup failed, dropping\n");
-        return XDP_DROP;
-    }
-    return action;
-}
-
-// End.DT6: Decapsulation with IPv6 table lookup
-// RFC 8986 Section 4.7
-// Like End.DX6 but uses VRF-aware FIB lookup via entry->vrf_ifindex
-static __always_inline int process_end_dt6(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    // 1. RFC 8986: SL must be 0 for DT6
-    if (srh->segments_left != 0) {
-        DEBUG_PRINT("End.DT6: SL != 0, passing\n");
-        return XDP_PASS;
-    }
-
-    // 2. Strip outer IPv6+SRH, expose inner IPv6
-    if (srv6_decap(ctx, srh, IPPROTO_IPV6) != 0) {
-        DEBUG_PRINT("End.DT6: Decapsulation failed\n");
-        return XDP_DROP;
-    }
-
-    // 3. Re-fetch pointers after adjust_head
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-
-    // 4. Validate Ethernet + IPv6 headers
-    struct ethhdr *eth = data;
-    if ((void *)(eth + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    struct ipv6hdr *inner_ip6h = (void *)(eth + 1);
-    if ((void *)(inner_ip6h + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    // 5. EtherType is already IPv6 (unchanged)
-
-    // 6. FIB lookup with VRF ifindex (0 falls back to ingress ifindex)
-    __u32 fib_ifindex = entry->vrf_ifindex ? entry->vrf_ifindex : ctx->ingress_ifindex;
-    DEBUG_PRINT("End.DT6: Decapsulated, FIB ifindex=%d\n", fib_ifindex);
-    STATS_INC(STATS_SRV6_END, 0);
-
-    int action = srv6_fib_redirect(ctx, inner_ip6h, eth, fib_ifindex);
-    // After decapsulation, must not return XDP_PASS
-    if (action == XDP_PASS) {
-        DEBUG_PRINT("End.DT6: FIB lookup failed, dropping\n");
-        return XDP_DROP;
-    }
-    return action;
-}
-
-// End.DT46: Decapsulation with IP (v4 or v6) table lookup
-// RFC 8986 Section 4.9
-// Detects inner protocol from SRH nexthdr, dispatches to DT4 or DT6
-static __always_inline int process_end_dt46(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    // Detect inner protocol and dispatch (SL check is done by callee)
-    __u8 inner_proto = srh->nexthdr;
-
-    if (inner_proto == IPPROTO_IPIP) {
-        DEBUG_PRINT("End.DT46: Inner protocol is IPv4, dispatching to DT4\n");
-        return process_end_dt4(ctx, ip6h, srh, entry);
-    } else if (inner_proto == IPPROTO_IPV6) {
-        DEBUG_PRINT("End.DT46: Inner protocol is IPv6, dispatching to DT6\n");
-        return process_end_dt6(ctx, ip6h, srh, entry);
-    }
-
-    DEBUG_PRINT("End.DT46: Unsupported inner protocol %d\n", inner_proto);
-    return XDP_DROP;
-}
-
-// Find peer_index by {bd_id, outer_src} via bd_peer_reverse_map.
-// O(1) single hash lookup instead of iterating up to MAX_BUM_NEXTHOPS entries.
-// Returns BD_PEER_INDEX_INVALID if not found.
-static __always_inline __u16 find_peer_index_by_src(
-    __u16 bd_id,
-    struct in6_addr *outer_src)
-{
-    struct bd_peer_reverse_key rk = { .bd_id = bd_id };
-    __builtin_memcpy(rk.src_addr, outer_src, sizeof(struct in6_addr));
-
-    struct bd_peer_reverse_val *rv = bpf_map_lookup_elem(&bd_peer_reverse_map, &rk);
-    if (!rv)
-        return BD_PEER_INDEX_INVALID;
-    return rv->index;
-}
-
-// End.DT2: Decapsulation with L2 table lookup (FDB) + remote MAC learning
-// RFC 8986 Section 4.11
-// Strips outer headers, learns inner src MAC as remote, then looks up
-// dst MAC in fdb_map for fast-path forwarding.
-// Unknown unicast (FDB miss) falls through to kernel bridge via XDP_PASS.
-static __always_inline int process_end_dt2(
-    struct xdp_md *ctx,
-    struct ipv6hdr *ip6h,
-    struct ipv6_sr_hdr *srh,
-    struct sid_function_entry *entry)
-{
-    // 1. SL must be 0
-    if (srh->segments_left != 0) {
-        DEBUG_PRINT("End.DT2: SL != 0, passing\n");
-        return XDP_PASS;
-    }
-
-    // 2. nexthdr must be IPPROTO_ETHERNET
-    if (srh->nexthdr != IPPROTO_ETHERNET) {
-        DEBUG_PRINT("End.DT2: nexthdr is not Ethernet (%d)\n", srh->nexthdr);
-        return XDP_DROP;
-    }
-
-    // 3. Save outer IPv6 source for remote MAC learning (before decap)
-    struct in6_addr outer_src;
-    __builtin_memcpy(&outer_src, &ip6h->saddr, sizeof(struct in6_addr));
-
-    // 4. Strip outer headers (Eth + IPv6 + SRH)
-    int strip_len = calc_decap_strip_len(srh);
-    if (bpf_xdp_adjust_head(ctx, strip_len)) {
-        DEBUG_PRINT("End.DT2: bpf_xdp_adjust_head failed\n");
-        return XDP_DROP;
-    }
-
-    // 5. Re-fetch pointers after adjust_head
-    void *data = (void *)(long)ctx->data;
-    void *data_end = (void *)(long)ctx->data_end;
-    struct ethhdr *inner_eth = data;
-    if ((void *)(inner_eth + 1) > data_end) {
-        return XDP_DROP;
-    }
-
-    // 6. Remote MAC learning: learn inner src MAC as remote entry
-    if (entry->bd_id != 0 && !(inner_eth->h_source[0] & 0x01)) {
-        __u16 peer_idx = find_peer_index_by_src(entry->bd_id, &outer_src);
-        if (peer_idx != BD_PEER_INDEX_INVALID) {
-            struct fdb_key learn_key = { .bd_id = entry->bd_id };
-            __builtin_memcpy(learn_key.mac, inner_eth->h_source, ETH_ALEN);
-
-            struct fdb_entry *existing = bpf_map_lookup_elem(&fdb_map, &learn_key);
-            if (!existing || !existing->is_remote ||
-                existing->peer_index != peer_idx) {
-                struct fdb_entry learn_val = {
-                    .is_remote = 1,
-                    .peer_index = peer_idx,
-                    .bd_id = entry->bd_id,
-                };
-                bpf_map_update_elem(&fdb_map, &learn_key, &learn_val, BPF_ANY);
-                DEBUG_PRINT("End.DT2: Learned remote MAC, peer_index=%d\n", peer_idx);
-            }
-        }
-    }
-
-    // 7. Build FDB lookup key for dst MAC (bd_id + MAC)
-    struct fdb_key key = { .bd_id = entry->bd_id };
-    __builtin_memcpy(key.mac, inner_eth->h_dest, ETH_ALEN);
-
-    // 8. Lookup dst MAC in fdb_map
-    struct fdb_entry *fdb = bpf_map_lookup_elem(&fdb_map, &key);
-    if (!fdb) {
-        // FDB miss: redirect to bridge device for flooding.
-        // XDP_PASS would send to the uplink's kernel stack (not a bridge member),
-        // so the decapped L2 frame would be dropped.
-        if (entry->bridge_ifindex != 0) {
-            DEBUG_PRINT("End.DT2: FDB miss, redirect to bridge ifindex %d\n", entry->bridge_ifindex);
-            return bpf_redirect(entry->bridge_ifindex, 0);
-        }
-        DEBUG_PRINT("End.DT2: FDB miss, no bridge, passing to kernel\n");
-        return XDP_PASS;
-    }
-
-    // 9. Known unicast: fast-path redirect (local entries only)
-    if (!fdb->is_remote) {
-        DEBUG_PRINT("End.DT2: FDB hit local, redirect to ifindex %d\n", fdb->oif);
-        STATS_INC(STATS_SRV6_END, 0);
-        return bpf_redirect(fdb->oif, 0);
-    }
-
-    // Remote FDB hit on receiving PE — should not happen in normal operation
-    // (packet arrived from remote, dst is also remote = routing loop)
-    DEBUG_PRINT("End.DT2: FDB hit remote, passing to kernel\n");
-    return XDP_PASS;
-}
-
-#endif // SRV6_ENDPOINT_H
